+++ workalendar/tests/test_brazil.py
@@ -676,6 +676,15 @@ class BrazilCuiabaCityTest(BrazilMatoGrossoTest):
         # Fixed days
         # Aniversário de Cuiabá
         self.assertIn(date(2017, 4, 8), holidays)
+        self.assertIn(date(2017, 4, 14), holidays)  # good_friday
+        self.assertIn(date(2017, 4, 16), holidays)  # easter_sunday
+        self.assertIn(date(2017, 6, 15), holidays)  # corpus_christi
+
+    def test_good_friday_label(self):
+        holidays = self.cal.holidays(2017)
+        holidays_dict = dict(holidays)
+        good_friday_label = holidays_dict[date(2017, 4, 14)]
+        self.assertEqual(good_friday_label, "Sexta-feira da Paixão")
 
 
 class BrazilBelemCityTest(BrazilParaTest):